requirements.txt
Pillow==8.4.0
requests==2.24.0
selenium==4.0.0
tqdm==4.50.2
